# backend/internal/docs/swagger.yaml
basePath: /api/
definitions:
  _ResponseWithData:
    description: Общий ответ success/error, содержащий произвольные данные.
    properties:
      data:
        description: Объект полезной нагрузки
      status:
        description: Результат запроса
        type: string
    type: object
  _ResponseWithMessage:
    description: Общий простой ответ, который передает только понятное для человека
      сообщение.
    properties:
      message:
        description: Человеко-читаемое сообщение
        type: string
      status:
        description: Результат запроса
        type: string
    type: object
  APIKey:
    description: Модель API ключа пользователя.
    properties:
      created_at:
        description: Дата создания
        example: "2025-10-25T13:40:00Z"
        format: date-time
        type: string
      expires_at:
        description: Дата истечения (если задана)
        example: "2025-11-25T13:40:00Z"
        format: date-time
        type: string
      id:
        description: ID ключа
        example: 7b2aab2e-4d1f-45b5-90c5-4d5d4db5ef11
        type: string
      name:
        description: Название ключа (для понимания)
        example: mobile_app
        type: string
      revoked:
        description: Отозван ли ключ
        example: false
        type: boolean
      user_id:
        description: ID пользователя
        example: 1a2b3c4d-5678-90ab-cdef-1234567890ab
        type: string
    type: object
  APIKeyCreateRequest:
    description: Запрос на создание API ключа.
    properties:
      name:
        description: Имя ключа
        example: partner_api
        type: string
      ttl_hours:
        description: Время жизни в часах (0 — бессрочно)
        example: 720
        type: integer
    required:
    - name
    type: object
  APIKeyCreateResponse:
    description: Ответ при создании API ключа.
    properties:
      api_key:
        description: Секретный API ключ (показывается один раз)
        example: KJHsT9W-2oP3sA1Q-LzM8fD4eC
        type: string
    type: object
  APIKeyListResponse:
    description: Ответ со списком ключей пользователя.
    properties:
      keys:
        description: Список всех активных ключей
        items:
          $ref: '#/definitions/APIKey'
        type: array
    type: object
  APIKeyRevokeRequest:
    description: Запрос на отзыв API ключа.
    properties:
      id:
        description: ID ключа
        example: 7b2aab2e-4d1f-45b5-90c5-4d5d4db5ef11
        type: string
    required:
    - id
    type: object
  Article:
    description: модель статьи.
    properties:
      content_en:
        description: Содержимое статьи на английском
        example: Example
        type: string
      content_ru:
        description: Содержимое статьи на русском
        example: Пример
        type: string
      createdAt:
        description: Timestamp создания аккаунта
        example: "2006-01-02T15:04:05Z"
        format: date-time
        type: string
      id:
        description: ID статьи (UUID)
        example: b4b03119-1290-44bc-b599-6a5e91d6611f
        type: string
      title_en:
        description: Заголовок статьи на английском языке
        example: Example
        type: string
      title_ru:
        description: Заголовок статьи на русском языке
        example: Пример
        type: string
      updatedAt:
        description: Timestamp последнего обновления аккаунта
        example: "2006-01-02T15:04:05Z"
        format: date-time
        type: string
    required:
    - content_en
    - content_ru
    - createdAt
    - id
    - title_en
    - title_ru
    - updatedAt
    type: object
  ArticleCreateRequest:
    description: Данные, передаваемые для создания статьи.
    properties:
      content_en:
        description: Содержимое статьи на английском
        type: string
      content_ru:
        description: Содержимое статьи на русском
        type: string
      title_en:
        description: Заголовок статьи на английском языке
        type: string
      title_ru:
        description: Заголовок статьи на русском языке
        type: string
    required:
    - content_en
    - content_ru
    - title_en
    - title_ru
    type: object
...
  hackathon-back_internal_model.SearchResult:
    properties:
      article:
        $ref: '#/definitions/Article'
      highlight:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Флоу авторизации: сначала пользователь регистрируется, в ответе получает модельку своего user и токен.
    Токен + код с почты он отправляет на ручку confirm, если что-то идёт не так (токен истёк, не пришёл код на почту), то запрос нужно отправить на ручку resend-confirmation.
    Далее уже можно авторизоваться, login/refresh/test-login выставляет в cookie access и refresh токены, фронтенду, ничего с ними делать не нужно, они сами по себе живут в браузере и отправляются при каждом запросе.
    Специально для мобильного приложения при login/refresh/test-login токены дублируются в теле ответа.
    При запросе к защищённым ручкам API мобильному приложению необходимо выставить заголовок Authorization: Bearer *access_token*.
    При refresh мобильное приложение передаёт refresh токен в теле запроса.
  title: Hackathon API
  version: 0.1.0
paths:
  /article:
    post:
      consumes:
      - application/json
      description: Создать статью.
      parameters:
      - description: Данные для создания статьи
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/ArticleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/Article'
              type: object
        "400":
          description: Invalid body param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to get article
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Создать статью.
      tags:
      - Articles
  /article/{article_id}:
    delete:
      description: Удалить статью по ID.
      parameters:
      - description: Article UUID
        in: path
        name: article_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Invalid path param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to delete article
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Удалить статью.
      tags:
      - Articles
    get:
      description: Получить статью по ID.
      parameters:
      - description: Article UUID
        in: path
        name: article_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/Article'
              type: object
        "400":
          description: Invalid path param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to get article
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Получить статью по ID.
      tags:
      - Articles
    patch:
      consumes:
      - application/json
      description: Частично обновить статью. Некоторые поля могут быть пустыми, обновляются
        только те, которые не пустые.
      parameters:
      - description: Article UUID
        in: path
        name: article_id
        required: true
        type: string
      - description: Поля для обновления статьи
        in: body
        name: updateArticle
        required: true
        schema:
          $ref: '#/definitions/ArticleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Invalid path param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to update article
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Частично обновить статью.
      tags:
      - Articles
  /article/search:
    get:
      description: Полнотекстовый поиск по статьям.
      parameters:
      - description: Строка поиска
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/hackathon-back_internal_model.SearchResult'
                  type: array
              type: object
        "400":
          description: Invalid query param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to get articles
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Поиск статей по содержанию.
      tags:
      - Articles
  /auth/api-key:
    post:
      consumes:
      - application/json
      description: Создаёт новый API ключ для текущего пользователя. Ключ можно использовать
        для аутентификации через заголовок `X-API-Key`.
      parameters:
      - description: Данные для создания API ключа
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/APIKeyCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ключ успешно создан
          schema:
            $ref: '#/definitions/APIKeyCreateResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на стороне сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание нового API ключа
      tags:
      - API Keys
  /auth/api-key/list:
    get:
      description: Возвращает все активные и неотозванные ключи текущего пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Список ключей
          schema:
            $ref: '#/definitions/APIKeyListResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на стороне сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка всех API ключей пользователя
      tags:
      - API Keys
  /auth/api-key/revoke:
    post:
      consumes:
      - application/json
      description: Делает указанный API ключ недействительным. После этого он не может
        быть использован для аутентификации.
      parameters:
      - description: ID ключа для отзыва
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/APIKeyRevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ключ успешно отозван
        "400":
          description: Некорректный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Ключ не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на стороне сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отзыв (деактивация) API ключа
      tags:
      - API Keys
...
swagger: "2.0"